feat: Integrate redux on Cart Empty page (#22)
* feat: Integrate redux on Cart Empty page

* refactor: move Cart sub-components to components folder

* refactor: export Cart
