--- pages/cart/index.js
@@ -1,14 +1,0 @@
-import Content from './components/CartContent/CartContent';
-import CartImage from './components/CartImage/CartImage';
-import Empty from './components/CartContent/CartEmpty';
-
-const Index = () => {
-  return (
-    <>
-      <CartImage />
-      <Content />
-    </>
-  );
-};
-
-export default Index;
